mcp_client_app/requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
azure-core